refactor : package 분리, domain 객체가 가지고있는 view 객체 개선을 위한 구조 분리
