+++ src/main/java/study/domain/AutoRacingParticipant.java
@@ -1,4 +1,6 @@
-package study;
+package study.domain;
+
+import study.utils.RandomNumberGenerator;
 
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -13,31 +15,12 @@ public class AutoRacingParticipant {
         autoRacingMembers.addAll(racingCar);
     }
 
-    public void doGame(int numberOfGame) {
-        for (int i = 0; i < numberOfGame; i++) {
-            gameEachRound();
-            ResultView.lineChangeEachRoundFinish();
-        }
-        ResultView.printName(getAllMaxMoveCountName());
-    }
-
-    private void gameEachRound() {
-
+    public List<RacingCar> gameEachRound() {
         for (RacingCar racingCar : autoRacingMembers) {
-            ResultView.printName(racingCar.getName());
             int randomNumber = RandomNumberGenerator.generateRandomNumber();
             racingCar.moveOrStop(randomNumber);
-            ResultView.printResultEachRound(racingCar.getMoveCount());
         }
-
-    }
-
-    private int getMaxCountValue() {
-        return autoRacingMembers.stream()
-                .max(Comparator
-                        .comparing(RacingCar::getMoveCount))
-                .orElseThrow(NoSuchElementException::new)
-                .getMoveCount();
+        return autoRacingMembers;
     }
 
     public List<String> getAllMaxMoveCountName() {
@@ -48,4 +31,14 @@ public class AutoRacingParticipant {
                 .collect(Collectors
                         .toList());
     }
+
+    private int getMaxCountValue() {
+        return autoRacingMembers.stream()
+                .max(Comparator
+                        .comparing(RacingCar::getMoveCount))
+                .orElseThrow(NoSuchElementException::new)
+                .getMoveCount();
+    }
+
+
 }
