+++ src/main/java/study/domain/MoveCount.java
@@ -1,12 +1,24 @@
-package study;
+package study.domain;
 
 import java.util.Objects;
 
 public class MoveCount {
     private int moveCount;
 
+    public MoveCount() {
+        new MoveCount(0);
+    }
+
+    public MoveCount(int number) {
+        this.moveCount = number;
+    }
+
     public void moveForward() {
         moveCount++;
+    }
+
+    public boolean checkIsSame(int maxCountValue) {
+        return moveCount == maxCountValue;
     }
 
     @Override
@@ -22,19 +34,8 @@ public class MoveCount {
         return Objects.hash(moveCount);
     }
 
-    public MoveCount(int number) {
-        this.moveCount = number;
-    }
-
-    public MoveCount() {
-        new MoveCount(0);
-    }
-
-    public boolean checkIsSame(int maxCountValue) {
-        return moveCount == maxCountValue;
-    }
-
     public int getCount() {
         return this.moveCount;
     }
+
 }
