+++ src/main/java/study/domain/Name.java
@@ -1,10 +1,18 @@
-package study;
+package study.domain;
 
 import java.util.Objects;
 
 public class Name {
-    private String name;
+    private final String name;
     private static final int MAX_LENGTH_OF_NAME = 5;
+
+    public Name(String name) {
+        if (name.length() >= MAX_LENGTH_OF_NAME) {
+            throw new IllegalArgumentException("사용자의 이름의 길이는 5이상으로 작성할수 없습니다.");
+        }
+        this.name = name;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
@@ -18,12 +26,6 @@ public class Name {
         return Objects.hash(name);
     }
 
-    public Name(String name) {
-        if (name.length() >= MAX_LENGTH_OF_NAME) {
-            throw new IllegalArgumentException("사용자의 이름의 길이는 5이상으로 작성할수 없습니다.");
-        }
-        this.name = name;
-    }
 
     public String getName() {
         return this.name;
