+++ src/main/java/study/utils/MoveChecker.java
@@ -1,13 +1,10 @@
-package study;
+package study.utils;
 
 public class MoveChecker {
     private static final int MOVE_BOUNDARY = 4;
 
     public static boolean canMoveForward(int number) {
-        if (number >= MOVE_BOUNDARY) {
-            return true;
-        }
-        return false;
+        return number >= MOVE_BOUNDARY;
     }
 
 }
