+++ src/main/java/study/utils/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package study;
+package study.utils;
 
 import java.util.Random;
