+++ src/main/java/study/utils/StringSplitter.java
@@ -1,4 +1,4 @@
-package study;
+package study.utils;
 
 import java.util.Arrays;
 import java.util.List;
