+++ src/main/java/study/view/ResultView.java
@@ -1,7 +1,8 @@
-package study;
+package study.view;
+
+import study.domain.RacingCar;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class ResultView {
     private static final String GUIDE_FOR_BEFORE_GAME_START = "실행결과=";
@@ -12,17 +13,30 @@ public class ResultView {
         System.out.println(GUIDE_FOR_BEFORE_GAME_START);
     }
 
+    public static void printResultEachRound(List<RacingCar> racingCarList) {
+        for (RacingCar racingCar : racingCarList) {
+            printName(racingCar.getName());
+            printAfterName();
+            printResultEachRound(racingCar.getMoveCount());
+        }
+        lineChangeEachRoundFinish();
+    }
+
     public static void printName(String name) {
         System.out.print(name);
     }
 
+    private static void printAfterName() {
+        System.out.print(" : ");
+    }
 
-    public static void printResultEachRound(int moveCount) {
+    private static void printResultEachRound(int moveCount) {
         for (int i = 0; i < moveCount; i++) {
             System.out.print(MOVE_COUNT_STRING);
         }
         System.out.println();
     }
+
 
     public static void lineChangeEachRoundFinish() {
         System.out.println();
