+++ src/test/java/study/utils/CarFactoryTest.java
@@ -1,8 +1,9 @@
-package study;
+package study.utils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.domain.RacingCar;
 
 import java.util.ArrayList;
 import java.util.List;
