+++ src/test/java/study/utils/RandomNumberGeneratorTest.java
@@ -1,10 +1,9 @@
-package study;
+package study.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class RandomNumberGeneratorTest {
     @Test
