+++ src/test/java/study/utils/StringSplitterTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
